// src/resources/calendar/types.ts
export type Weekday = "SU" | "MO" | "TU" | "WE" | "TH" | "FR" | "SA";
// ...
export interface ZonedFields {
  year: number; // full year
  month: number; // 1..12
  day: number; // 1..31
  hour: number; // 0..23
  minute: number; // 0..59
  second: number; // 0..59
  ms?: number; // 0..999
}